refactor: put events-related shared code in root _utils
